feat(vendor): align price lists with Figma
Port the vendor price-list surface to the marketplace offer-scoped model
(mirroring the admin MER-160 sync) and extract the shared, sdk-free core into
@mercurjs/dashboard-shared so both panels use one source of truth.

- Offer-scoped prices: every override carries rules.offer_id so a seller's
  override cannot leak onto another seller's offer of the same variant
- List: add Type column, Type/Status/Created/Updated filters, Title-ascending
  default sort, "No price lists yet" empty state, drop the header subtitle
- Create wizard: Offers picker + offer-keyed Prices grid (Offer/Variant + SKU),
  Status radio (Draft default), clean date labels, Customer availability builder
- Detail: Offers section (offer-grouped, Add / Edit prices / Remove), Configuration
  card (dates only), separate Customer availability section + edit drawer
- Edit form: Status radio + optional description; offer-scoped add/edit-prices
- Offers rows link to /offers/:id (admin appends ?seller_id); vendor offer
  fetches drop seller.id/seller.name (all vendor offers are seller-scoped)
- Success toasts match Figma verbatim ("Price list was successfully updated.")
- Backend: extend VendorGetPriceListsParams with type/created_at/updated_at
- Shared: new @mercurjs/dashboard-shared/price-lists (enums, offer-grid builder,
  extractPricesFromOffers/getPriceListStatus); admin re-exports from it

Adds integration-tests/http/price-lists/vendor/price-list-offers.spec.ts asserting
offer-scoped price isolation, and a dev-stability fix for the
@mercurjs/vendor/extension-targets Vite resolution flake (default exports + alias).
