+++ packages/admin/src/pages/price-lists/common/constants.ts
@@ -1,20 +1,5 @@
-/**
- * Re-implementation of enum from `@medusajs/medusa` as it cannot be imported
- */
-export enum PriceListStatus {
-  ACTIVE = "active",
-  DRAFT = "draft",
-}
-
-export enum PriceListDateStatus {
-  SCHEDULED = "scheduled",
-  EXPIRED = "expired",
-}
-
-/**
- * Re-implementation of enum from `@medusajs/medusa` as it cannot be imported
- */
-export enum PriceListType {
-  SALE = "sale",
-  OVERRIDE = "override",
-}
+export {
+  PriceListDateStatus,
+  PriceListStatus,
+  PriceListType,
+} from "@mercurjs/dashboard-shared"
