+++ packages/dashboard-shared/src/index.ts
@@ -24,3 +24,4 @@ export {
 export * from "./components"
 export * from "./hooks"
 export * from "./extensions"
+export * from "./price-lists"
