+++ packages/dashboard-shared/src/price-lists/constants.ts
@@ -1,0 +1,18 @@
+/**
+ * Re-implementation of enums from `@medusajs/medusa` as they cannot be imported
+ * directly. Shared by the admin and vendor price-list surfaces.
+ */
+export enum PriceListStatus {
+  ACTIVE = "active",
+  DRAFT = "draft",
+}
+
+export enum PriceListDateStatus {
+  SCHEDULED = "scheduled",
+  EXPIRED = "expired",
+}
+
+export enum PriceListType {
+  SALE = "sale",
+  OVERRIDE = "override",
+}
